+++ src/utils/hooks/usePlayTime.js
@@ -1,11 +1,11 @@
 // import { useMemo } from 'react'
-import { useTrackPlayerProgress } from '@/plugins/player/hook'
+import { useProgress } from '@/plugins/player/utils'
 import { formatPlayTime2 } from '@/utils'
 // import { useGetter } from '@/store'
 // import { STATE_PLAYING, STATE_BUFFERING } from 'react-native-track-player'
 
 export default () => {
-  const { position, bufferedPosition, duration } = useTrackPlayerProgress(250)
+  const { position, buffered, duration } = useProgress(250)
   // const isGettingUrl = useGetter('player', 'isGettingUrl')
 
   return {
@@ -13,7 +13,7 @@ export default () => {
     maxTimeStr: formatPlayTime2(duration),
     time: position,
     duration,
-    bufferedProgress: duration ? bufferedPosition / duration * 100 : 100,
+    bufferedProgress: duration ? buffered / duration * 100 : 100,
     progress: duration ? position / duration * 100 : 0,
   }
 }